- Updated PersonResource to mirror SQl DB responses and PUT requires whole entity - Updated Person to have a Request inner class with validation - Updated tests for Person and PersonResource - tests passing - Created separate middleware for DBException and ViolationException mappers - TODO: Now Dao tests are failing, need to debug

+++ src/main/java/com/hellowiz/api/db/PersonDAO.java
@@ -21,9 +21,10 @@ public interface PersonDAO {
     int insert(@Bind("name") String name, @Bind("email") String email);
 
     @SqlUpdate("UPDATE persons SET name = :name, email = :email WHERE id = :id")
-    int updateById(@Bind("id") long id, @Bind("name") String name, @Bind("email") String email);
+    int updateById(@Bind("id") int id, @Bind("name") String name, @Bind("email") String email);
+
     @SqlQuery("SELECT * FROM persons WHERE id = :id")
-    Person findById(@Bind("id") long id);
+    Person findById(@Bind("id") int id);
 
     @SqlQuery("SELECT * FROM persons WHERE email = :email")
     Person findByEmail(@Bind("email") String email);
